Project2/pyproject.toml
[project]
name = "project2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio>=4.9.0",
    "distro>=1.9.0",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "jiter>=0.10.0",
    "openai>=1.86.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "flask",
    "pymupdf",
]
